fix: resolve the compile errors reported by CI
- MainActivity: `video/*` inside the KDoc closed the comment early, which also
  broke FileOpsService's reference to MainActivity.EXTRA_OPEN_TRANSFERS
- CoilSetup: Coil 2 keeps Fetcher/FetchResult/SourceResult in coil.fetch, and
  the image loader takes CoroutineDispatchers rather than ExecutorServices
- SettingsRepository: DataStore's [] operator needs MutablePreferences + the
  `set` extension import
- Favorites/RecentStore: fromJson sits in a companion object, so use the plain
  org.json accessors instead of JsonStore's protected helpers
- SafDocProvider: DocumentsContract.moveDocument requires a non-null parent
- VolumeMonitor: StorageVolumeCallback is an abstract class, not a SAM type
- PlayerViewModel: Media3 exposes no MediaItem.subtitleConfigurations, so track
  the configurations per playlist index
- PlayerScreen: Media3 has no RESIZE_MODE_STRETCH (labels adjusted to match)
- Dialogs: Row takes horizontalArrangement
- SettingsScreen: ClosedFloatingPointRange.start, not startInclusive
- Bitmaps: frameScore mixed Double statistics with a Float result

+++ app/src/main/java/com/usbmediaexplorer/util/Bitmaps.kt
@@ -155,9 +155,9 @@ object Bitmaps {
 
         // Brightness term peaks around mid-grey.
         val brightnessScore = when {
-            mean < 8 -> mean / 8f * 0.15f
-            mean > 245 -> ((255 - mean) / 10f).coerceAtLeast(0f) * 0.15f
-            else -> 0.55f + (1f - kotlin.math.abs(mean - 128f) / 128f) * 0.25f
+            mean < 8.0 -> (mean / 8.0 * 0.15).toFloat()
+            mean > 245.0 -> (((255.0 - mean) / 10.0).coerceAtLeast(0.0) * 0.15).toFloat()
+            else -> (0.55 + (1.0 - kotlin.math.abs(mean - 128.0) / 128.0) * 0.25).toFloat()
         }
         val detailScore = (std / 74.0).coerceIn(0.0, 1.0).toFloat()
         val colourScore = (colourfulness / 40.0).coerceIn(0.0, 1.0).toFloat()
